requirements.txt
rich>=13.0.0
tabulate>=0.9.0
python-dateutil>=2.8.0
pytest>=7.0.0
